Seção 9 - DTOs e Data Anotations

+++ Back/src/ProEventos.API/Controllers/EventosController.cs
@@ -1,14 +1,9 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Linq;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Mvc;
-using Microsoft.Extensions.Logging;
-using ProEventos.Persistence;
-using ProEventos.Domain;
-using ProEventos.Persistence.Contextos;
 using ProEventos.Application.Contratos;
 using Microsoft.AspNetCore.Http;
+using ProEventos.Application.Dtos;
 
 namespace ProEventos.API.Controllers
 {
@@ -31,7 +26,11 @@ namespace ProEventos.API.Controllers
             try
             {
                 var eventos = await _eventoService.GetAllEventoAsync(true);
-                if(eventos == null) return NotFound("Nenhum evento encontrado.");
+                if(eventos == null) return NoContent();
+                
+                
+
+
                 return Ok(eventos);
             }
             catch (Exception ex)
@@ -47,7 +46,7 @@ namespace ProEventos.API.Controllers
             try
             {
                 var evento = await _eventoService.GetEventoByIdAsync(id,true);
-                if(evento == null) return NotFound("Evento por Id encontrado.");
+                if(evento == null) return NoContent();
 
                 return Ok(evento);
             }
@@ -65,7 +64,7 @@ namespace ProEventos.API.Controllers
             try
             {
                 var evento = await _eventoService.GetAllEventosByTemaAsync(tema,true);
-                if(evento == null) return NotFound("Eventos por tema não encontrados.");
+                if(evento == null) return NoContent();
 
                 return Ok(evento);
             }
@@ -78,7 +77,7 @@ namespace ProEventos.API.Controllers
 
         }
         [HttpPost]
-        public async Task<IActionResult> Post(Evento model)
+        public async Task<IActionResult> Post(EventoDto model)
         {
             try
             {
@@ -96,7 +95,7 @@ namespace ProEventos.API.Controllers
         }
 
         [HttpPut("{id}")]
-        public async Task<IActionResult> Put(int id,Evento model)
+        public async Task<IActionResult> Put(int id,EventoDto model)
         {
            try
             {
@@ -118,9 +117,12 @@ namespace ProEventos.API.Controllers
         {
            try
             {
+                var evento = await _eventoService.GetEventoByIdAsync(id);
+                if(evento ==null) return NoContent();
+
                 return await _eventoService.DeleteEvento(id)?
                 Ok("Deletado"):
-                BadRequest("Evento não deletado");
+                throw new Exception("Ocorreu um problema não especicico ao tentar deletar o Evento");
             }
             catch (Exception ex)
             {
